requirements.txt
Django>=4.0.3,<4.1
djangorestframework>=3.14.0,<3.15
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.26.2,<0.27
Pillow>=9.5.0,<9.6.0
stripe>=5.4.0,<5.5.0
django-phonenumber-field>=7.1.0,<7.2.0
phonenumberslite>=8.13.14,<8.14.0